val struct

+++ binary_type.h
@@ -49,8 +49,8 @@ int read_mysql_tiny(const char *buf);
 int read_mysql_short(const char *buf);
 int read_mysql_longlong(const char *buf);
 char *read_mysql_newdecimal(const char *buf, const char *meta, int *cursor);
-int read_mysql_float(const char *buf);
-int read_mysql_double(const char *buf);
+float read_mysql_float(const char *buf);
+double read_mysql_double(const char *buf);
 
 int read_mysql_varchar(const char *buf, const char *meta);
